@@ -1,76 +1,39 @@
 {
-  "_args": [
-    [
-      {
-        "raw": "word-wrap@^1.0.3",
-        "scope": null,
-        "escapedName": "word-wrap",
-        "name": "word-wrap",
-        "rawSpec": "^1.0.3",
-        "spec": ">=1.0.3 <2.0.0",
-        "type": "range"
-      },
-      "D:\\dev\\react-native-nativebutton\\node_modules\\cz-conventional-changelog"
-    ]
+  "name": "word-wrap",
+  "description": "Wrap words to a specified length.",
+  "version": "1.2.3",
+  "homepage": "https://github.com/jonschlinkert/word-wrap",
+  "author": "Jon Schlinkert (https://github.com/jonschlinkert)",
+  "contributors": [
+    "Danilo Sampaio <danilo.sampaio@gmail.com> (localhost:8080)",
+    "Fede Ramirez <i@2fd.me> (https://2fd.github.io)",
+    "Joe Hildebrand <joe-github@cursive.net> (https://twitter.com/hildjj)",
+    "Jon Schlinkert <jon.schlinkert@sellside.com> (http://twitter.com/jonschlinkert)",
+    "Todd Kennedy (https://tck.io)",
+    "Waldemar Reusch (https://github.com/lordvlad)",
+    "Wolfgang Faust (http://www.linestarve.com)",
+    "Zach Hale <zachhale@gmail.com> (http://zachhale.com)"
   ],
-  "_from": "word-wrap@>=1.0.3 <2.0.0",
-  "_id": "word-wrap@1.2.1",
-  "_inCache": true,
-  "_location": "/word-wrap",
-  "_nodeVersion": "6.9.2",
-  "_npmOperationalInternal": {
-    "host": "packages-18-east.internal.npmjs.com",
-    "tmp": "tmp/word-wrap-1.2.1.tgz_1485796135235_0.5473448638804257"
-  },
-  "_npmUser": {
-    "name": "jonschlinkert",
-    "email": "github@sellside.com"
-  },
-  "_npmVersion": "3.10.9",
-  "_phantomChildren": {},
-  "_requested": {
-    "raw": "word-wrap@^1.0.3",
-    "scope": null,
-    "escapedName": "word-wrap",
-    "name": "word-wrap",
-    "rawSpec": "^1.0.3",
-    "spec": ">=1.0.3 <2.0.0",
-    "type": "range"
-  },
-  "_requiredBy": [
-    "/cz-conventional-changelog"
-  ],
-  "_resolved": "https://registry.npmjs.org/word-wrap/-/word-wrap-1.2.1.tgz",
-  "_shasum": "248f459b465d179a17bc407c854d3151d07e45d8",
-  "_shrinkwrap": null,
-  "_spec": "word-wrap@^1.0.3",
-  "_where": "D:\\dev\\react-native-nativebutton\\node_modules\\cz-conventional-changelog",
-  "author": {
-    "name": "Jon Schlinkert",
-    "url": "https://github.com/jonschlinkert"
-  },
+  "repository": "jonschlinkert/word-wrap",
   "bugs": {
     "url": "https://github.com/jonschlinkert/word-wrap/issues"
   },
-  "dependencies": {},
-  "description": "Wrap words to a specified length.",
+  "license": "MIT",
+  "files": [
+    "index.js",
+    "index.d.ts"
+  ],
+  "main": "index.js",
+  "engines": {
+    "node": ">=0.10.0"
+  },
+  "scripts": {
+    "test": "mocha"
+  },
   "devDependencies": {
     "gulp-format-md": "^0.1.11",
     "mocha": "^3.2.0"
   },
-  "directories": {},
-  "dist": {
-    "shasum": "248f459b465d179a17bc407c854d3151d07e45d8",
-    "tarball": "https://registry.npmjs.org/word-wrap/-/word-wrap-1.2.1.tgz"
-  },
-  "engines": {
-    "node": ">=0.8"
-  },
-  "files": [
-    "index.js"
-  ],
-  "gitHead": "2ae779e2089e83d9e32b7c05382c4a1ac2b1d5c1",
-  "homepage": "https://github.com/jonschlinkert/word-wrap",
   "keywords": [
     "break",
     "carriage",
@@ -85,28 +48,6 @@
     "words",
     "wrap"
   ],
-  "license": "MIT",
-  "main": "index.js",
-  "maintainers": [
-    {
-      "name": "doowb",
-      "email": "brian.woodward@gmail.com"
-    },
-    {
-      "name": "jonschlinkert",
-      "email": "github@sellside.com"
-    }
-  ],
-  "name": "word-wrap",
-  "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jonschlinkert/word-wrap.git"
-  },
-  "scripts": {
-    "test": "mocha"
-  },
   "typings": "index.d.ts",
   "verb": {
     "toc": false,
@@ -132,6 +73,5 @@
       "verb",
       "verb-generate-readme"
     ]
-  },
-  "version": "1.2.1"
+  }
 }
